+++ tests/compilation_failures/expressions/late_bound_owned_to_mutable copy.rs
@@ -1,0 +1,11 @@
+use preinterpret::*;
+
+fn main() {
+    let _ = preinterpret!{
+        #(
+            let a = "a";
+            "b".swap(a);
+            a
+        )
+    };
+}