+++ tests/compilation_failures/expressions/owned_to_mutable.rs
@@ -1,0 +1,11 @@
+use preinterpret::*;
+
+fn main() {
+    let _ = preinterpret!{
+        #(
+            let a = "a";
+            a.swap("b");
+            a
+        )
+    };
+}